+++ packages/permit/CHANGELOG.md
@@ -1,6 +1,12 @@
 # @zap-studio/permit
 
-All notable changes to this package are documented in this file.
+## 0.2.1
+
+### Patch Changes
+
+- 5ea3d3b: Updated dependencies.
+  - @zap-studio/validation@0.3.1
+
 
 ## 0.2.0